refactor(lib)!: refactor all traits in lib
This huge refactor rework all the trait system used in the library to
push types to the surface instead of having one context: Swap.

This allows more flexibility but refactor is not completly done, some
trait should be removed and some should be split, e.g. crypto::Sign, to
allow less generics in protocol functions.

This refactor allows proper implementation/derivation on messages that
are used in node and passed over the network.