refactor: reverse trait system (#255)
* refactor: remove serde feature gate

first step to refactor serde in this library.

* refactor: remove unnecessary trait bonds on FeeStrat.

* fix: remove optional serde to fix compilation

* refactor: simplify crypto types

Simplify trait bounds on TaggedElement and take direct generic on
AccordantKeys and AccordandKeySet instead of nested generic.

* feat: add more primitive derives

* doc: fix missing link

* refactor: rename accordant keys fields

* feat: add serde derive on remaining types

* refactor: script types to disentangle generics

* refactor: new protocol mod and message generics

Move protocol_message into the new protocol module. Modify the generic
argument to split the context into multiple generic. Add more derives on
types.

This commit prepare the removing of bundles from core. They are in fact
messages created inside node already. It also prepare the migration of
execution logic on Alice and Bob into the new protocol module.

* refactor: migrate protocol execution logic

Alice and Bob logic is migrated into protocol module.

* refactor(lib)!: refactor all traits in lib

This huge refactor rework all the trait system used in the library to
push types to the surface instead of having one context: Swap.

This allows more flexibility but refactor is not completly done, some
trait should be removed and some should be split, e.g. crypto::Sign, to
allow less generics in protocol functions.

This refactor allows proper implementation/derivation on messages that
are used in node and passed over the network.

* re-enable taproot module

* ci: fix test unknown serde feat

* chore: remove commented bundles

* chore: remove commented code

* fix: reimplement Encodable/Decodable on Bob